// Frontend/src/generated/model/cartItem.ts
/**
 * SportsHub
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CartItem { 
    id?: any;
    userId?: any;
    productId?: any;
    quantity?: any;
    size?: any;
    createdAt?: any;
}